requirements.txt
# Backend Python dependencies for Finalv3.2
# Pin ranges for stability across Windows environments with YOLOv8 + DeepSort + FastAPI
# Tip: For GPU acceleration, install a CUDA-specific torch build from PyTorch instructions.

# Web API
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0

# Computer vision and tracking
ultralytics>=8.0,<9.0
# Torch is required by Ultralytics; including explicit pins for reliability
torch>=2.1,<3.0
torchvision>=0.16,<1.0
opencv-python>=4.8,<5.0
numpy>=1.24,<3.0

deep-sort-realtime>=1.3,<2.0

# Database (PostgreSQL). Use the binary wheel for easier Windows installs
psycopg2-binary>=2.9,<3.0

# Utilities
python-dateutil>=2.8,<3.0

# Optional (uncomment if you use environment files)
# python-dotenv>=1.0,<2.0

# Testing
pytest>=7.0,<9.0
# Pin httpx below 0.28 for compatibility with starlette TestClient in FastAPI 0.110
httpx>=0.24,<0.28
